requirements_dev.txt
click
flake8
fpdf2
imageio
pylint
pytest
pytest-coverage
